Enforce structural output; clean streamlit app script
