+++ pyproject.toml
@@ -7,6 +7,7 @@ requires-python = ">=3.12"
 dependencies = [
     "google-genai>=1.22.0",
     "groq>=0.29.0",
+    "instructor>=1.9.2",
     "langchain-openai>=0.3.27",
     "langsmith>=0.4.4",
     "openai>=1.92.3",
